backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.129.1",
    "httpx>=0.28.1",
    "icalendar>=7.0.1",
    "langchain-openai>=1.1.10",
    "langchain>=0.0.332",
    "google-generativeai>=0.8.6,<0.9.0",
    "langsmith>=0.4.0",
    "langgraph>=1.0.9",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.22",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]
